Changes to support the newer avr-1 devices (starting with attiny814) by tweaks90

+++ uart.h
@@ -235,7 +235,7 @@ were adapted from the Arduino HardwareSerial.h library by Tim Sharpe on
    @param   baudrate Specify baudrate using macro UART_BAUD_SELECT()
    @return  none
 */
-extern void uart0_init(uint16_t baudrate);
+extern void uart0_init(uint32_t baudrate);
 
 
 /**
